+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - 添加 API `/api/storage/setLocalStorageVal` | Add API `/api/storage/setLocalStorageVal`
 - 为 API `/api/file/getFile` 的 `blob` 响应结果添加 `content-type` | Add `content-type` to the `blob` response result of API `/api/file/getFile`
 - [#1](https://github.com/siyuan-community/siyuan-sdk/issues/1) 添加 API `/api/filetree/getIDsByHPath` | Add API `/api/filetree/getIDsByHPath`
+- 调整 API `/api/snippet/getSnippet` | Adjust API `/api/snippet/getSnippet`
 
 ## v0.3.3 / 2023-10-24